Refactor project helper code
Projects are now synced when editing, on a 5 minute throttle
